pyproject.toml
[tool.poetry]
name = "starter-bot"
version = "0.1.0"
description = ""
authors = ["nandhagk <nandhagopi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
